requirements.txt
slack-bolt>=1.18.0
anthropic>=0.40.0
mcp>=1.0.0
PyGithub>=2.1.1
jira>=3.8.0
python-dotenv>=1.0.0
httpx>=0.27.0
starlette>=0.41.0
uvicorn>=0.32.0
gitpython>=3.1.43
anyio>=4.6.0
